Refactor symbolic state value printing
This cleans up and refactors the iteration over state variable
valuation/value pairs via StateValues. In particular StateValuesMTBDD
learns how to do non-sparse printing.

Furthermore, it
* introduces generic "iterate" and "iterateFiltered"
  methods for iterating over the state/values, calling a
  StateAndValueConsumer object
* provides a StateAndValueConsumer implementation (StateAndValuePrinter)
  that supports printing in the various output styles supported by PRISM
* cleans up the convenience "print" method using default methods in StateValues
* activates the MTBDD engine for one -exportvector based test

Note: Previously, in the StateValuesMTBDD iteration steps, the int[]
array with the current values represented the increments to be added
to the low values of state variables, with the addition
  value = low + a[i]
happening on output. Now, we initialise the array with low, i.e.,
the values in the array correspond to the actual variable values.