GUIMultiModelHandler: Inhibit creation of WaitParseThread during GUI startup
Previously, during the constructor of GUIMultiModelHandler, a WaitParseThread
was created. If the startup of the other plugins takes longer than the configured delay
(default = 1s, but can be much shorter via settings), the attempted parse can generate
events that reach other plugins/components that are not yet fully initialised, leading
to NullPointerExceptions, etc.

Now, we inhibit the creation of the WaitParseThread until the startup has
completed.

Fixes #111.