Make ModulesFileModelGenerator and associated classes generic.
The "associated classes" are Updater, TransitionList, Choice
and ChoiceListFlexi.

This is inline with the previous conversion of ModelGenerator to generic form.
As there, existing use of ModulesFileModelGenerator should be largely
unaffected, since it is set up to work using doubles by default.
In particular, the existing constructors continue to work. However,
for compatibility going forward, it is preferable to now use the static
ModulesFileModelGenerator.create(...) method instead.

In another related change, the check that distributions sum to 1
when constructing models with the explicit engine now uses the method
PrismUtils.doublesAreEqual, which is what is implemented in the default
Evaluator for doubles. This uses relative error with epsilon = 1e-12,
which is a tighter check than before, but should be fine to allow
the check to function in the presence of round-off errors.
This also means that PRISM_SUM_ROUND_OFF from PrismSettings is
now ignored by the explicit engine.