PrismPaths helper: Resolve file system specifications, relative to a model or properties file.
This helper can be used to resolve paths given in PRISM properties,
e.g., the location of an automaton file in a path expression.

If the path is relative, we will resolve it relative to
 1) the properties file location (if there is one)
 2) the model file location (if there is one)
 3) the working directory.

Additionally, it's also possible to specify paths using the following three macros:
 $MODEL_DIR/filename
 $PROPERTY_DIR/filename
 $WORKING_DIR/filename
which resolve relative to the directory where the model file or the property file is
located or the working directory.