ModulesFile, PropertiesFile: Store location (path to file), make accessible
If we know the filename of a ModulesFile or PropertiesFile, we store
it and make it accessible.  Note that we don't necessarily have a
filename, e.g., when dealing with model code from the GUI that has not
yet been saved.

Additionally, for symbolic/explicit StateModelChecker, we provide
access to the ModulesFile / PropertiesFile stored in the model
checker.