prism-auto: --log-subdirs option to create benchmark subdirectories for logs.
When using the -l (--log) option and also the -a (--args) option,
subirectories are created for each entry in the args file.
This makes it easier to process the log files afterwards.

Relatedly, if the directory specified in the -l switch, or the required
subdirectories, do not exist, there is now no error and they are created.