prism-auto: --args-list option (command-line version of --args)
Like --args <file>, this allow a list of lists of arguments to be provided,
with each element of the outer list representing a separate run of PRISM
to be performed on all benchmarks. Here, these can be provided directly
on the command-line, as a comma-separated list, rather than creating a file.

For example:

 prism-auto ... --args-list '-gs,-jac,-jor -omega 0.9'