Add support for model checking partially observable MDPs (POMDPs).
This is a rebased version of the code previously published in the prism-ext repo,
with various bits of code refactored, tidied and improved.

Model construction and model checking support is added to the explicit engine.
The PRISM modelling language is extended with the 'pomdp' model keyword
and observables...endobservables blocks to indicate observable variables.