Add -arg switch to set current working dir in command line and GUI.
This sets the working directory natively, which is used to read or
write files specified with relative locations. For example:

  prism -dir /some/dir model.prism model.props

is the same as:

  prism /some/dir/model.prism /some/dir/model.props

Exported files also go to the same directory.

The same applies when passing model/property files to the GUI via
the command-line and when specifying export filenames as options.
The GUI file chooser is also initialised to the -dir location.