Remove initial variable value info from VarList.
VarList should focus on the range, structure, etc. of the variables.
The initial values of the variables for a model are stored separately.

This info was only accessed from Modules2MTBDD,
which now extracts initial variable values itself,
and uses the VarList.encodeToInt helper method.
Checking that bounded integer vars are in range is now moved there.