Enable (explicit engine) model checking of LTSs.
There is no explicit model type keyword added to the parser.
Instead, LTSs are auto-detected by the absence of probabilities.

ConstructModel is extended to build LTSs. Model checking (CTL+LTL)
can be done out of the box using the existing NonProbModelChecker class.

A few MDP regression tests are clarified to make it clear that the
models are MDPs, not LTSs.