HOAF2DA: Support HOA input without a number-of-states header
The States header in HOA automata is optional. Previously, we aborted
if it was missing.  As we can allocate states in the generated DA
on-the-fly, to support parsing a HOA automaton with absent State
header, we just have to dynamically allocated states and track that
all states that are used somewhere are actually defined.