Stricter check in POMDPSimple that actions match for observations.
It requires that the available actions are the same and appear in
the same order (in terms of choice indexing) in each equivalent state.

This always happens for PRISM models built via ConstructModel since
the model construction considers actions one by one in the same
order in each state (and actions must be unique for POMDPs).