GUI simulator fix: add properties at the correct time.
The correct time is after path creation, but before initialisation.
In fact, a previous commit has made this unnecessary, by allowing
paths to be added slightly later, but this is still the right thing to do.

Simultaneously, change the interface to SimulatorEngine.loadPath(),
requiring createNewPath() to be called separately beforehand.
This is currently only used when loading paths from the GUI.