Complete re-write of explicit.StateValues.
The main goal is to allow storage of any Object; previously only ints,
doubles and booleans were allowed. As before, a StateValues has a Type
and the Objects stored should be of the appropriate type to match that.
This means that values obtained from the vector can always be safely
cast to the expected Object type.

Internally, the data is stored either as a BitSet, for Booleans, or
an array of Objects, for other types. The specific arrays of primitives
used previously are gone for now. This is slightly less efficient, but
efficiency is not a high priority for these classes.

All the type-specific operations, such as applying binary/unary
operators, functions, etc. are removed, with the logic having been
moved to classes such as ExpressionBinaryOp or ExpressionFunc.
Instead generic functions to apply predicates/functions are provided.

Specific functions for Boolean functions (and, or, iff, etc.) are
retained since they may be performed more efficiently using BitSet
methods, but these methods are (mostly) not called currently.