Implement direct conversion of double to BigRational
The previous implementation first converted the double
value into a string which is then parsed.

The new algorithm avoids this detour and requires no canceling.
Also, it is not affected by the precision of the actual string
representation.

It decomposes a double into its exponential representation:

  value = integer * 2^exp

Then, the equivalent fraction is

  value = integer / 2^(-exp)

Determining the exponent by repeated multiplication with 2
yields the smallest exponent satisfying the above equations.
Hence, no canceling is required.