Allow importing model.all without import switch.
The standard CLI expects an import switch to import
a model as matrix from text files.
If the filname is given with the extension .all,
PRISM finds the relevant files automatically.

This mechanism be used to simplify the CLI for .all:

  prism -importmodel model.all properties.props ...

becomes

  prism model.all properties.props ...

This allows the user to use models given in the PRISM
languages or as explicit representation interchangably.