Add option to export model and props labels to same file
Both -exportmodel and -exportlabels have an option to export
labels from a model and properties file to the same file.
The default behaviour only exports labels defined in the model.

Note: This changes the default from the previous version which
always exported labels from a properties file if present.

Fixes #179