Extend ModelGenerator to support real-time models (e.g., PTAs).
States are passed to a ModelGenerator for exploration as usual.
But, when evaluating which choices/transitions are available, the values of
any clock variables should be ignored, and a choice is considered
available if it could be for any clock variable valuation. For real-time
models, method getChoiceClockGuard() can be called to get the clock guard
(as an Expression) for any enabled choices.

ModelGenerator also has another new method getClockInvariant(), which returns
the clock invariant, again as an Expression over clock variables, for the
state currently being explored.

Since the value of clock variables in the current state is ignored,
it is acceptable for the values of these variables to be null.

On the other hand, the computeTransitionTarget() method does provide
the new values of clock variables in the resulting State object,
therefore providing information about which clocks are reset.

ModulesFileModelGenerator implements this for PRISM models.

We explicitly disable real-time simulation for now, since this
still requires further work.