PTA construction from a PRISM model is now done via a model generator.
This removes previous restrictions on (non digital clocks) PTA engines,
such as the use of global variables or access to non-local variables,
and will allow easier addition of new language features in future.

For now, the PTA construction only works on PRISM language models
(i.e, a ModulesFile object), but this is just because the labels stored
there might be needed for model checking. Later, PTA construction can
be generalised to work from any ModelGenerator.

The old code in Modules2PTA (which did a syntactic exploration of each
module separately) is no longer needed, but not deleted.

The new code is in ConstructPTA and bears some similarity to the code
in ConstructModel for normal explicit state engine model construction.
One difference at the code level is that "locations" in the PTA data structure
are now stored as State objects with null stored for clock variables,
rather than as substates. This is to future-proof against later changes.
This should mostly be seamless, but error messages change slightly, e.g.:

Error: Timelock (no transitions) in PTA at location (s=0,x=?,y=?).