ModulesFile and PropertiesFile use EvaluateContext for setting/storing constants.
The preferred setSomeUndefinedConstants() method, to set constants, now takes an
EvaluateContext. The EvaluateContext is available via getUndefinedEvaluateContext(),
similar to before. The old set methods still work, but are deprecated.

Locally, we also store the evaluated values for _all_ constants in an
EvaluateContext. More importantly, this also retains the evaluation mode.
This firms up the expected usage of ModulesFile and PropertiesFile in that
setSomeUndefinedConstants() specifies both the necessary constant values
and the evaluation mode, and that these remain the same until called again.
They are available by calling getEvaluateContext().

These changes are also applied to the ModelInfo interface.

In line with the above, ModulesFile.getDefaultInitialState() takes no arguments.

In a related change, the setUndefinedConstants() methods, which threw an
exception if any constants were not defined, are deprecated. They were not in the
ModelInfo interface, and are not really used, e.g., because some verification
methods such as parametric might leave constants undefined, and properties file
constants are usually dealt with on a per-property basis. It is preferred to call
setSomeUndefinedConstants() instead, and problems with missing constants will be
picked up at a later point.