Add Type.defaultDeclarationType() method.
This is used to allow ModelInfo.getVarDeclarationType() to be
implemented automatically in some common cases. The logic from
there is pushed into individual Type classes.

We also add DeclarationDoubleUnbounded, which is a default
for the double type. Currently no double variables are allowed
but this will later be useful for storing constant types.