Proper support for SimulatorEngine following a strategy.
prism.Prism now retains a Strategy if generated/returned during model checking,
accessible via getStrategy(). There is a new method loadStrategyIntoSimulator()
to pass the current strategy to the simulator.

SimulatorEngine has a new mechanism for loading a strategy, as a StrategyGenerator,
which takes care of providing a strategy and mapping it to the State view of a model.
It is also optional whether it is enforced or not, during automatic path  generation,
via method ​setStrategyEnforced(). The state/decision of the strategy can  be queried
for the current state: isTransitionEnabledByStrategy(), getStrategyDecisionString().

Path classes (optionally) store strategy info (current decision and memory) in states.