Refactor ASTElement deepCopy() into clone() and DeepCopy visitor.
This commit (combined with the previous two) separates deep-copying
of an AST element into two steps:

1. Creating a shallow copy of the AST node
2. Recursively cloning the fields and child nodes using a visitor DeepCopy

This enables the programmer to create both shallow copies (aka clones in
Java terminology) and deep copies. The visitor DeepCopy provides a hook
into the copying process. This can be useful, e.g., to modify the copy on
the fly or cache copied elements. It also provides a cleaner interface
since traversal and copying logic is better separated.

To simplify copying child nodes, DeepCopy provides the utility methods

To adapt existing and new subclasses of ASTElement:

1. Override #clone such that (a) the type of the callee is returned and
   (b) all collection fields are cloned too. See Command#clone() and
   ConstantList#clone for examples.

2. Override the abstract method #deepCopy(DeepCopy) such that it (a)
   returns the type of the callee and (b) deep copies its child nodes
   and fields using the utility methods of the provided visitor. See
   Command#deepCopy and ConstantList#deepCopy for examples. These
   methods are shorter and cleaner than the old #deepCopy method.

Closes #217.