API change/fix: loadModelFromExplicitFiles needs rewards file _list_.
Attempt to retain old method with just one File fails, becase the
methods are ambiguous in the common case where null is passed.

Use Collections.singletonList(stateRewardsFile) when there is only one.

Also fix a bug so that loadModelFromExplicitFiles can be passed null
(that was currently never happening anyway when called from PrismCL).