Refactor and extend strategy export options.
Create StrategyExportOptions class to manage options.

This is now passed to the various Strategy.exportX() methods
(which is a change in the interface, effecting both external
calls and any other implementing classes). The exportStrategy
methods in Prism change similarly.

Using this, the -exportstrat switch, managed by PrismCL,
is extended with several new options: mode, states, obs.
We also add an alias type=model for type=induced.

These options are not yet all fully supported but obs
does now provides the choice to merge POMDP states or not.