Exact model checking (via parametric) integrated properly in Prism.
In exact mode, model construction and storage is now handled in prism.Prism in
the same way as for other engines. Model checking is now triggered by the usual
modelCheck() method and modelCheckExact() is now deprecated.

A few changes are made in the parametric model checking code to align the way
that results are stored and displayed. Unsupported aspects of model checking are
now detected and reported more cleanly.

Model export is now also supported for exact mode.