Code for import of labels and initial states moved to PrismExplicitImporter.
Now called by both ExplicitFiles2Model and ExplicitFiles2MTBDD.

We retain explicit.StateModelChecker.loadLabelsFile, now reimplemented
using PrismExplicitImporter, since this is sometimes used elsewhere.

Note that the decision that an absence of a labels file means the initial
state defaults to 0 is now specified by the importer, not the model builder.