Check for transition reward consistency when importing from .trew files.
It is assumed that rewards are the same for all successor states from
the same state-choice pair, since we don't support state-choice-state
rewards yet. So the explicit engine now checks that this is the case
when importing a .trew file.

In fact, the symbolic engines _can_ import these (and a check is actually
a bit tricky), so we don't check for the symbolic engines.