Refactor handling of imported transition reward indexing.
For Markov chains, the importer is now responsible for providing
transition rewards indexed either by the transition's offset
within the state or by the successor state. The former is used
by the explicit engine, the latter by the symbolic engine.

The user calls setTransitionRewardIndexing to specify which
indexing is required.

There is also a new method setModel in ExplicitModelImporter
which can be used to provide access to the model in case needed
during transition reward processing. Notably this is used when
extracting rewards from a .trew file for use in the explicit
engine (which requires conversion of state->offset indexing).
This logic is now moved into PrismExplicitImporter.