Update action name storage in explicit models.
New method findActionsUsed() in prism.Model interface,
used in default implementation of getActions().
The execution of this can be optimised by implementing
onlyNullActionUsed() and when all actions are null.

New class ActionList to encapsulate info about action names.
This is stored in model base classes, notably ModelExplicit,
meaning that looking up action indices is much faster.

For now, the usage of ActionList is that it is initialised
to be empty, and the actual list is computed lazily when needed.
Changes to the model that affect actions should call the method
markNeedsRecomputing() in case the list has been computed.