ModelInfo.getActions() supplies all actions used.
This includes null for unlabelled choices/transitions (as for Model).

This will allow aligning actions in a built model and its source,
facilitating the use of action indices for efficiency.

For now, this is optional: returning null indicates that the
information is unavailable.

ModulesFile supports this. BasicModelInfo is able to store/provide the info
but, for now, explicit model import does not use this because it cannot
provide the action list efficiently without searching all transitions.