Explicit model find/fix-deadlocks methods update action list if needed.
(i.e. add null if an unlabelled self-loop was added to fix a deadlock)

Not actually needed if the action set is later computed on demand, since
the method to add a transition will take care of this, but this works
more smoothly when the assumed action set is provided up front with
ModelExplicit.setActions, as in e.g. ConstructModel.