Optionally, export/import initial state info via .tra file.
These are included as dummy transitions, with irrelevant fields
shown as "-", i.e., each initial state s is represented by a line
"- s -" or "- - s -" for Markov chains and MDPs, respectively.

If a .lab file is read from too, the initial states info is also
extracted from there and checked for consistency.

For now, this is only actually enabled for POMDPs,
where it is required to include observations for initial states.